+++ CHANGELOG.md
@@ -7,8 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.5.1-rc.1] - 2025-01-21
 
+### Added
+- Added a test for the function getAbsolutePath
+- Added a test for the function getGatheringRefs
+
 ### Fixed
 - Fixed an error in generating models according to the specification when the schema referred to the schema in another file. And the schema from the file referred to the schema inside the same file.
+
+### Updated
+- Moved the gatheringRefs function to a separate file
 
 ## [0.5.0-beta.0] - 2024-08-01